Simplify design of snapshotting
