+++ src/MessageRepository.php
@@ -11,4 +11,6 @@ interface MessageRepository
     public function persist(Message ...$messages);
 
     public function retrieveAll(AggregateRootId $id): Generator;
+
+    public function retrieveAllAfterVersion(AggregateRootId $id, int $aggregateRootVersion): Generator;
 }
